Feat & Refactor ( Eslint, config, name updates + )
