--- components/Card.js
@@ -1,13 +1,0 @@
-export default function Card () {
-  return (
-    <div>
-      <div>
-        <img src="https://images.pexels.com/photos/4381392/pexels-photo-4381392.jpeg?auto=compress&cs=tinysrgb&dpr=1&w=500" alt="tailwind logo" class="rounded-xl" />
-      </div>
-      <div>
-        <h3>The Majestic and Wonderful Bahamas</h3>
-        <p>The best kept secret of The Bahamas is the country’s sheer size and diversity. With 16 major islands, The Bahamas is an unmatched destination</p>
-      </div>
-    </div>
-  )
-}
