--- components/Footer.js
@@ -1,7 +1,0 @@
-export default function Footer() {
-    return (
-        <footer>
-            <p>Copyright&copy; 2022</p>
-        </footer>
-    )
-}
