+++ components/Footer.jsx
@@ -1,0 +1,7 @@
+export default function Footer() {
+  return (
+    <footer>
+      <p>Copyright&copy; 2022</p>
+    </footer>
+  );
+}
