+++ next.config.js
@@ -1,3 +1,9 @@
 module.exports = {
   reactStrictMode: true,
+  images: {
+    domains: [
+      'images.pexels.com',
+      'via.placeholder.com'
+    ],
+  },
 }
