+++ package.json
@@ -25,8 +25,14 @@
     "@storybook/react": "^6.4.9",
     "autoprefixer": "^10.4.0",
     "babel-loader": "^8.2.3",
-    "eslint": "8.5.0",
-    "eslint-config-next": "12.0.7",
+    "eslint": "^8.5.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-next": "^12.0.7",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-import": "^2.25.3",
+    "eslint-plugin-jsx-a11y": "^6.5.1",
+    "eslint-plugin-react": "^7.28.0",
+    "eslint-plugin-react-hooks": "^4.3.0",
     "eslint-plugin-storybook": "^0.5.3"
   }
 }
