+++ pages/_app.jsx
@@ -1,9 +1,9 @@
-import Layout from '../components/layout'
+import Layout from '../components/Layout';
 
 export default function MyApp({ Component, pageProps }) {
   return (
     <Layout>
       <Component {...pageProps} />
     </Layout>
-  )
+  );
 }
