+++ pages/_document.jsx
@@ -1,12 +1,17 @@
-import Document, { Html, Head, Main, NextScript } from 'next/document'
+import Document, {
+  Html,
+  Head,
+  Main,
+  NextScript,
+} from 'next/document';
 
 class MyDocument extends Document {
   render() {
     return (
       <Html>
         <Head>
-          <link 
-            href="https://fonts.googleapis.com/css2?family=Playfair+Display&family=Roboto&display=swap" 
+          <link
+            href="https://fonts.googleapis.com/css2?family=Playfair+Display&family=Roboto&display=swap"
             rel="stylesheet"
           />
         </Head>
@@ -15,8 +20,8 @@ class MyDocument extends Document {
           <NextScript />
         </body>
       </Html>
-    )
+    );
   }
 }
 
-export default MyDocument
+export default MyDocument;
