+++ pages/archive.jsx
@@ -1,0 +1,5 @@
+export default function Archive() {
+  return (
+    <h1>Blog page</h1>
+  );
+}
