--- pages/contact.js
@@ -1,5 +1,0 @@
-export default function Contact() {
-    return (
-        <h1>Contact page</h1>
-    )
-}
