+++ pages/contact.jsx
@@ -1,0 +1,5 @@
+export default function Contact() {
+  return (
+    <h1>Contact page</h1>
+  );
+}
