+++ pages/index.jsx
@@ -2,11 +2,8 @@ import Card from '../components/Card';
 
 export default function Home() {
   return (
-    <div className="flex flex-wrap">
-      <Card />
-      <Card />
-      <Card />
+    <div>
       <Card />
     </div>
-  )
+  );
 }
