+++ pages/posts/first-post.jsx
@@ -1,0 +1,3 @@
+export default function FirstPost() {
+  return <h1>First Post</h1>;
+}
